# Fix not all text being changed to white when timing tools heatmap enabled