@Joeclinton1 2025-09-21
@Samq64 Works now in my testing. Thanks for catching that edge case.

I also made it so that it will update the heatmap even when the project isn't running if it finds any block has been modified since last update. Because during my testing for this bug it was annoying me that I'd click the green flag with heatmap and profiling enabled and it wouldn't colour the blocks until I actually enabled the heatmap.